+++ soxs/background/diffuse.py
@@ -22,7 +22,7 @@ XSPEC model used to create the "default" foreground spectrum
           7.3e-07
 
 XSPEC model used to create the "lem" foreground spectrum
-  model  apec + wabs*(apec+apec)
+  model  apec + tbabs*(apec+apec)
             0.099
                 1
                 0
@@ -43,11 +43,12 @@ def make_frgnd_spectrum(arf, rmf):
     bkgnd_nH = float(soxs_cfg.get("soxs", "bkgnd_nH"))
     absorb_model = soxs_cfg.get("soxs", "bkgnd_absorb_model")
     frgnd_spec_model = soxs_cfg.get("soxs", "frgnd_spec_model")
+    frgnd_velocity = float(soxs_cfg.get("soxs", "frgnd_velocity"))
     agen = ApecGenerator(rmf.ebins[0], rmf.ebins[-1], rmf.n_e,
-                         broadening=False)
-    spec = agen.get_spectrum(0.225, 1.0, 0.0, 7.3e-7)
+                         broadening=True)
+    spec = agen.get_spectrum(0.225, 1.0, 0.0, 7.3e-7, velocity=frgnd_velocity)
     if frgnd_spec_model == "halosat":
-        spec += agen.get_spectrum(0.7, 1.0, 0.0, 8.76e-8)
+        spec += agen.get_spectrum(0.7, 1.0, 0.0, 8.76e-8, velocity=frgnd_velocity)
     spec.apply_foreground_absorption(bkgnd_nH, model=absorb_model)
     spec += agen.get_spectrum(0.099, 1.0, 0.0, 1.7e-6)
     spec.restrict_within_band(emin=0.1)